[Refactor/#5] Activity -> Navigation 수정
