+++ app/src/main/java/org/sopt/at/history/HistoryScreen.kt
@@ -1,0 +1,25 @@
+package org.sopt.at.history
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.navigation.NavController
+import org.sopt.at.utils.BottomNavigation
+
+@Composable
+fun HistoryScreen(navController: NavController){
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(color = Color.Black)
+    ){
+        Text(
+            text = "History Screen"
+        )
+        BottomNavigation(navController = navController)
+    }
+}
