--- app/src/main/java/org/sopt/at/home/DramaViewModel.kt
@@ -1,19 +1,0 @@
-package org.sopt.at.home
-
-import androidx.lifecycle.ViewModel
-import org.sopt.at.R
-
-class DramaViewModel : ViewModel() {
-    val dramaItems = mutableListOf(
-        R.drawable.drama1_poster_image,
-        R.drawable.drama2_poster_image,
-        R.drawable.drama3_poster_image,
-        R.drawable.drama4_poster_image,
-        R.drawable.drama5_poster_image,
-        R.drawable.drama6_poster_image,
-        R.drawable.drama7_poster_image,
-        R.drawable.drama8_poster_image,
-        R.drawable.drama9_poster_image,
-        R.drawable.drama10_poster_image
-    )
-}
