--- app/src/main/java/org/sopt/at/home/ShortsView.kt
@@ -1,54 +1,0 @@
-package org.sopt.at.home
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.navigation.NavController
-
-@Composable
-fun ShortsView(navController: NavController){
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(color = Color.Black)
-    ){
-        BottomNavigation(navController = navController)
-    }
-}
-
-@Composable
-fun LiveView(navController: NavController){
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(color = Color.Black)
-    ){
-        BottomNavigation(navController = navController)
-    }
-}
-
-@Composable
-fun SearchView(navController: NavController){
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(color = Color.Black)
-    ){
-        BottomNavigation(navController = navController)
-    }
-}
-
-@Composable
-fun HistoryView(navController: NavController){
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(color = Color.Black)
-    ){
-        BottomNavigation(navController = navController)
-    }
-}
