+++ app/src/main/java/org/sopt/at/shorts/ShortsScreen.kt
@@ -1,0 +1,25 @@
+package org.sopt.at.shorts
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.navigation.NavController
+import org.sopt.at.utils.BottomNavigation
+
+@Composable
+fun ShortsScreen(navController: NavController){
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(color = Color.Black)
+    ){
+        Text(
+            text = "Shorts Screen"
+        )
+        BottomNavigation(navController = navController)
+    }
+}
